refactor(engine): Move trigger input validation to dsl.validation
